Added two new files
I added a new way to do the log in (log_in_ver2.html) and the other file that goes with it (Congrats.html). I know its kind of ugly when run, but at least it works. I'm gonna try to find a better way to do it, so I didnt want to paste over the original html file. Take a look and let me know what you think.